refactor: make `manifest_from_server` internal
